requirements.txt
elevenlabs==0.2.27
langchain-groq
streamlit
load_dotenv
langchain